Store delays as uint32_t
Some delays are bigger than 2^16.
Perhaps, it makes more sense to not store them as milliseconds.

+++ slave/music.h
@@ -5,7 +5,7 @@
 
 #include <avr/pgmspace.h>extern PROGMEM const uint8_t title-screen_tones**;
 extern PROGMEM const uint8_t title-screen_events**;
-extern PROGMEM const uint16_t title-screen_delays**;
+extern PROGMEM const uint32_t title-screen_delays**;
 
 
 #endif
